Create `useCaptureWalletConnected` hook
Extracts the code from `useSaveConnectedAddressToStore` hook to a separate hook.
This way it will be easier to get the tbtc balance that we want to attach to
Wallet Connected event.
