+++ src/hooks/posthog/usePosthog.ts
@@ -2,6 +2,7 @@ import { useEffect } from "react"
 import { featureFlags } from "../../constants"
 import * as posthog from "../../posthog"
 import { useCapturePageview } from "./useCapturePageview"
+import { useCaptureWalletConnected } from "./useCaptureWalletConnected"
 import { useIdentify } from "./useIdentify"
 
 export const usePosthog = () => {
@@ -13,4 +14,5 @@ export const usePosthog = () => {
 
   useCapturePageview()
   useIdentify()
+  useCaptureWalletConnected()
 }
